Extract random show axios call out to a manager
