+++ src/managers/jamchart.js
