--- utils/randomShow.js
@@ -1,7 +1,0 @@
-const request = require("postman-request");
-
-const url = `http://phish.in/api/v1/random-show`;
-
-const randomShow = async () => {};
-
-module.exports = randomShow;
